requirements.txt
discord.py>=2.3.2
requests>=2.31.0
beautifulsoup4>=4.12.2
selenium>=4.15.0
webdriver-manager>=4.0.1
python-dotenv>=1.0.0